requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
